backend/requirements.txt
fastapi==0.110.0
pydantic==2.6.4
pydantic-settings==2.2.1
uvicorn==0.29.0
python-json-logger==2.0.7
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
sqlalchemy==2.0.29
alembic==1.13.1
